Split CfaGeneration for supporting multiple procedures
